backend: add an email confirmation mechanism
Anytime a user register an email, we send a confirmation email to the
registered address. We use the built-in BackgrounTask to do that.
The confirm endpoint will change the value in db.

Note that in this pr the email send does not encode anything. We need to
secure the confirmation email.