alert_sys: improve parsing message
The position message are different than the alert message.
In the former, we need to handle GRB_CA and GRB_DEC that has three lines
each.

This code add a solution based on the fact that every message will
always be formated the same. In our case, this two fields has the
sequence of char ',\n', I remove the line feed and replace it with a
space. We can then parse as usual.

At the end I trim the empty space to make it easier to use.